Tighten admin app UX and remove dead UI plumbing
- Wire VITE_STORE_API_URL into admin frontend builds (prod + staging) so
  the orders page no longer 500s on missing env var. Staging looks up
  the store stack output via cloudformation so admin-only PRs don't
  require a store backend redeploy.
- Merge the activity feed into the dashboard and drop the separate
  /activity route and nav entry.
- Remove the "Good Roots Network" footer link (still reachable from the
  avatar menu) so the footer no longer wraps to a tall second row.
- Style the okra deny-reason dropdown to match the textarea, add a label
  gap, and rename "Okra queue" to "Okra" in the nav and section header.
- Drop "Products backed by Stripe" from the store catalog header — that
  was a leaky implementation detail.
- Drop the hardcoded "Handled in admin dashboard." note from seed
  request mark-as-done so the audit trail isn't full of boilerplate.
- Remove dead form state for statement_descriptor, nonprofit_program,
  and impact_summary from the StorePage form and request type — there
  were no inputs for them, so admins were silently round-tripping them.
- Remove the dead `tagline` prop from SiteFooter (declared required but
  never rendered) and clean up the call sites in admin/store/grn/web
  plus the orphan CSS rule.
- Shorten store nav labels: "Store catalog" → "Catalog", "Store orders"
  → "Orders" to match the new shorter Okra label.
