+++ apps/store/src/shell/AppShell.tsx
@@ -144,7 +144,6 @@ export function AppShell({ session, onSignIn, children }: AppShellProps) {
         <div className="store-shell-main__inner">{children}</div>
       </main>
       <SiteFooter
-        tagline="Growing food, sharing seeds, and helping more people feel at home on the land."
         meta={`${new Date().getFullYear()} Olivia's Garden Foundation. All rights reserved.`}
         links={footerLinks}
         socialLinks={[
