+++ apps/web/src/site/chrome.tsx
@@ -179,9 +179,6 @@ export function SiteFooter({
 
   return (
     <SharedSiteFooter
-      tagline={currentPage.path === '/donate'
-        ? 'Founded in love. Growing for community.'
-        : 'Growing food, sharing seeds, and helping more people feel at home on the land.'}
       meta={`${new Date().getFullYear()} Olivia's Garden Foundation. All rights reserved.`}
       links={footerLinks}
       legalLinks={legalFooterLinks}
