+++ packages/ui/src/styles.css
@@ -222,12 +222,6 @@ body {
     letter-spacing: 0.1em;
   }
 
-  .og-site-footer__tagline {
-    max-width: none;
-    font-size: 0.82rem;
-    line-height: 1.3;
-  }
-
   .og-site-footer__meta,
   .og-site-footer__social {
     font-size: 0.72rem;
